Fix for incorrect barcode generatior
The RGhost barcode generator doesn't work anymore. It generates
incorrect barcodes that are unreadable. I don't know why, I'm not
a ruby developer, I just know it does.

This change uses Barby, which generates proper barcodes, for all the
linear type of barcodes. If barby doesn't know how to generate that
type of barcode we fall back to the original (broken) RGhost method.

Ideally RGhost should be removed entirely but I just wanted to fix
the linear barcodes since those are the ones I use and I'm not a
ruby developer.